Added two PDF modification scripts.
removefirstpdfpage.py does what it says on the tin, it removes the
first page of a PDF. Some publishers add a disclaimer or something
else as a first page of a manuscript when you download it from their
site. This bothers me, the script thus uses GhostScript and deletes
the first page. Only tested on Scientific Linux 6.
shrinkpdf.py does what it says on the tin, it shrinks the size of a
PDF, which is nice if you want to send a huge file to a collaborator.
Based on a command from ehow.com it uses GhostScript to convert a PDF
to either "screen", "ebook" or "printer", which reduces the file size.