PaganinIterator is now less chatty and more informative
Updated PaganinIterator to actually abide to the 'verbose' flag when
reconstructing the slices, otherwise `gridrec` simply overwhelms the
command line with output. When option.verbose is off (default) we thus
pipe the output of the sinogram generation and reconstruction to
/dev/null.
Give out more helpful info where necessary and inform the user what to
do if he/she doesn't want to test.