Added script to plot Delta/Beta
Based on data downloaded from CXRO, you can plot Delta and/or Beta of
a given compound.
To do that, go to http://henke.lbl.gov/optical_constants/getdb2.html,
request a "Text File", save it as .dat somewhere and plot it with this
script.
Useful if you have to talk with a user requesting Paganin reconstruction
of his scans...