Refactoring PaganinIterator
Preparing for new pipeline and selective slice reconstruction.
Putting the necessary stuff into own functions to make code a bit
clearer.
Loops for iterating through Delta and Beta should be more clear now,
since the "work" stuff is made via functions.
Now I need to implement the "selective reconstruction", which means that
we only want to reconstruct a set of sinograms, so the whole thing goes
quicker.
But for this I need Fedes help.