PaganinIterator refactored (wating for Fede)
Cleared up help and streamlined output to console.
After finding out that 'subprocess' needs a list (Thanks Heiner!)
switching the 'verbose' logic to the call of subprocess.
Writing the commands to a log file, so they could be redone after
the fact.
Only give the Fiji-call when we're not testing and informing the user
that there's a log file to look at after we're done.