Adapting to SGE pipeline
To nicely construct the calls to prj2sinSGE we need to have more
parameters. These are either read from the LogFile (default) or input by
the user. LogFile reading thus expanded.
Calls to prj2sinSGE are constructed a bit differently.
Since SGE takes care of sheduling, we can simply submit everything and
wait for it to finish. Theoretically we can thus remove lots of the
logic about checking folders, etc.
The calls are probably not correct yet, since Fede explained them to me
in 10 minutes. This version was used for a very first test, the bottom
of the file still contains the old version...