Pretty output and final debugging (JobID)
Output to both log and console is now a bit prettier and meaningful. The
Job ID was not always reliably detected, sometimes we got back a string
with the location of the log-file folder. The script now looks if the
number it gets back is an 'int', otherwise it breaks the script and
let's the user retry. Cause of the problem is unknown, also to Fede :)