Further debugging and (hopefully final) corrections
Correcting typo which hindered reconstructions to be calculated, I
forgot to add the slice number to the location of the fltp folder...
Alain suggested to slice to ROI to an even number of files, this seems
to help with the script running through all of the tasks.
Save the commandline call to the Logfile, this can help with debugging.
Nicer job-names and nicer output to the console all-around.