Make RotationCenter to be half of cam-ROI if nothing found in log.
If you do not run sinooff_tomcat_j.py prior to the PaganinIterator, you do not have information on the RotCenter in the logfile.
In this case just set half of the camera window width to be the rotation center and let the user decide if that's good or not.