Generate output directories as workaround
Submitting a large number of jobs did not work, the SGE queue complained
about missing directories (even though we use --createMissing). After a
lengthy (failed) debug session with GRecoman-Master Goran we were not
able to fix the problem (submitting all command manually worked, just
not one after the other). The workaround is to manually generate the
directories at the beginning of the loop...

Also added a bit of distinction in the command line output with color
and bold. And added some comments and clarifications in the code.