Working version with writing to log files
This version works (including the input options).
We expect the first of the stacked scans, check for other ones present,
load the bottom image of stack N and compare it to $Percent images from
the top of stack N+1.
The results are displayed (and saved as PNG) and also saved to a log
file in the rec folder.
This should help analyzing the "problem" that the overlap doesn't seem
to be consistent (seen at beamtime for user  e15171 on Oct 27, 2014).

Still missing is the reading of TIFF files and some memory
optimizations.