StackedScanOverlapFinder for all reconstructions
The stacked scan overlap finder now works for all kinds of
reconstructions (8bit, 16bit and DMP).
We achieved this by loading the EPD module (line 24), which loads a more
modern version of matplotlib.
We normalize the 8 and 16 bit images prior to comparison, making for a
more robust overlap detection.
Also making the whole file `pep8' compliant.